requirements.txt
attrs==22.2.0
boto3==1.26.84
botocore==1.29.84
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.0
iniconfig==2.0.0
jmespath==1.0.1
markdown-it-py==2.2.0
mdurl==0.1.2
packaging==23.0
pluggy==1.0.0
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
rich==13.3.2
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
yapf==0.32.0
